# Speedup query parsing