Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speedup query parsing

Detailed description / Documentation draft:

I've taken a long query, obfuscated it and made it even longer (while still valid) to have someting long to test.

* Results of the added perf test:

```
python3.8 docker/test/performance-comparison/perf.py tests/performance/explain_ast.xml --runs 100
```

MASTER:
```
query   0       explain_ast.query0.run114       0       0.06646919250488281
query   0       explain_ast.query0.run115       0       0.06641864776611328
client-time     0       8.104475416999776       8.0503511428833
profile-total   0
stage   run     8.191   8.227
stage   drop-2  0.000   8.227
```

Changes:
```
query   0       explain_ast.query0.run159       0       0.04686236381530762
query   0       explain_ast.query0.run160       0       0.04838275909423828
client-time     0       8.08941406699887        8.014005661010742
profile-total   0
stage   run     8.154   8.189
stage   drop-2  0.000   8.189
```

In 8 seconds it went from 115 to 160 queries, so about 1.4x as fast.

* Per commit with clickhouse-benchmark:
  * MASTER: localhost:9000, queries 500, QPS: **15.070**, RPS: 276426.933, MiB/s: 23.821, result RPS: 276426.933, result MiB/s: 23.821.
  * SET TO VECTOR: localhost:9000, queries 500, QPS: **18.057**, RPS: 331221.219, MiB/s: 28.543, result RPS: 331221.219, result MiB/s: 28.543.
  * TOKENITERATOR INLINE: localhost:9000, queries 500, QPS: **19.067**, RPS: 349753.045, MiB/s: 30.140, result RPS: 349753.045, result MiB/s: 30.140.
  * EXPECTED INLINE: localhost:9000, queries 500, QPS: **20.165**, RPS: 369883.637, MiB/s: 31.875, result RPS: 369883.637, result MiB/s: 31.875.
  * WRAP INLINE: localhost:9000, queries 500, QPS: **20.462**, RPS: 375336.799, MiB/s: 32.345, result RPS: 375336.799, result MiB/s: 32.345.

Tests with full release build.