@Algunenano 2021-11-29
Comparing between set and vector + find:
* set: localhost:9000, queries 500, **QPS: 17.141**, RPS: 314417.897, MiB/s: 27.095, result RPS: 314417.897, result MiB/s: 27.095.
* vector: localhost:9000, queries 500, **QPS: 20.758,** RPS: 380758.426, MiB/s: 32.812, result RPS: 380758.426, result MiB/s: 32.812.


Funnily enough, the vector + find is faster than before. 2 guesses: I simplified the 2 if's to reduce the number of branches and having a smaller vector makes it faster to copy the class when needed. Pushing it.

In terms of output, both discard duplicates. There is a change in output ordered by whatever order the set did (pointer comparison) and now they are output in order of insertion, so you get first the first thing CH would parse.


Before:

```
Syntax error: failed at position 19 (end of query):

Select a as b FROM 

Expected one of: VIEW, SELECT subquery, compound identifier, identifier, element of expression with optional alias, list of elements, function, table, table function, subquery or list of joined tables, table or subquery or table function
```

Now:
```
Syntax error: failed at position 19 (end of query):

Select a as b FROM 

Expected one of: table, table function, subquery or list of joined tables, table or subquery or table function, element of expression with optional alias, SELECT subquery, function, VIEW, identifier, compound identifier, list of elements
```