@nikitamikhaylov 2021-11-30
You can also try to replace `std::vector` to `boost::small_vector` in other places and check the performance