@Algunenano 2021-12-01
I've done a quick test with `boost::container::small_vector` with 4, 6 and 8 elements but the difference is smaller than the runs deviation so I can't say it improves things.